requirements.txt
pandas
numpy<1.17
keras
matplotlib
h5py
tensorflow
opencv-python
pillow